# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Final Project Backend API",
    "version": "0.1.0",
    "description": "API documentation for mobile and web clients. Most protected endpoints require a Supabase access token via Authorization: Bearer <token>."
  },
  "servers": [
    {
      "url": "https://www.medi-buddy.xyz",
      "description": "Production Server (HTTPS)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "AuthV2LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "AuthV2TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserAccount"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
...
      "AuthV2OtpRequestBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "AuthV2OtpVerifyRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "AuthV2OtpVerifyResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "isNewUser": {
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserAccount"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "isNewUser",
          "user"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "PublicUserAccount": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "supabaseUserId": {
            "type": "string",
            "nullable": true
          },
          "provider": {
            "type": "string",
            "nullable": true,
            "enum": [
              "email",
              "google",
              "email,google",
              null
            ]
          },
          "role": {
            "type": "string",
            "enum": [
              "User",
              "Admin",
              "SuperAdmin"
            ]
          },
          "tutorialDone": {
            "type": "boolean"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "email",
          "role",
          "tutorialDone",
          "createdAt"
        ]
      },
...
      "AdminDashboardUsageItem": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer"
          },
          "accountLabel": {
            "type": "string"
          },
          "patientCount": {
            "type": "integer"
          },
          "medicationLogCount": {
            "type": "integer"
          }
        },
        "required": [
          "accountId",
          "accountLabel",
          "patientCount",
          "medicationLogCount"
        ]
      },
      "AdminDashboardUsageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminDashboardUsageItem"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/auth/v2/login": {
      "post": {
        "tags": [
          "Auth (V2)"
        ],
        "summary": "Login with email/password",
        "description": "Authenticate user and return access/refresh tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthV2LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthV2TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/v2/otp/request": {
      "post": {
        "tags": [
          "Auth (V2)"
        ],
        "summary": "Request OTP code",
        "description": "Send a 6-digit verification code to the provided email address. Rate-limited to one request per minute per email.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthV2OtpRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests (cooldown active)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/v2/otp/verify": {
      "post": {
        "tags": [
          "Auth (V2)"
        ],
        "summary": "Verify OTP and login/register",
        "description": "Verify the 6-digit code. If the user exists, logs them in. If not, auto-creates a new account. Returns JWT tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthV2OtpVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verified — tokens issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthV2OtpVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/v2/users/delete": {
      "delete": {
        "tags": [
          "Admin - Users"
        ],
        "summary": "Bulk delete users strictly (V2)",
        "description": "Permanently delete multiple users and all their data (profiles, logs, relations). Requires 'CONFIRM' string.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userIds": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "Array of User IDs to delete"
                  },
                  "confirm": {
                    "type": "string",
                    "enum": [
                      "CONFIRM"
                    ],
                    "description": "Must be 'CONFIRM'"
                  }
                },
                "required": [
                  "userIds",
                  "confirm"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Users deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid confirmation or request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (e.g., trying to delete SuperAdmin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/v1/dashboard/usage": {
      "get": {
        "tags": [
          "Admin - Dashboard"
        ],
        "summary": "Get dashboard usage stats",
        "description": "Get usage stats per account (profile count and medication log count). You can filter by date range using fromDate / toDate.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (YYYY-MM-DD). Optional."
          },
          {
            "name": "toDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (YYYY-MM-DD). Optional."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDashboardUsageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}